添加SsqSchedule文件,生成双色球数据

+++ lottery-web/src/main/java/top/duanyd/lottery/dao/impl/DataURLDaoImpl.java
@@ -41,7 +41,7 @@ public class DataURLDaoImpl implements IDataURLDao {
                 dataURLEntity.setLotteryCode(resultSet.getString("lottery_code"));
                 dataURLEntity.setUrl(resultSet.getString("url"));
                 dataURLEntity.setParam(resultSet.getString("param"));
-                dataURLEntity.setStatus(resultSet.getByte("status"));
+                dataURLEntity.setStatus(resultSet.getInt("status"));
                 dataURLEntity.setCreateTime(resultSet.getTimestamp("create_time"));
                 dataURLEntity.setUpdateTime(resultSet.getTimestamp("update_time"));
                 dataURLEntity.setRemark(resultSet.getString("remark"));
